Slurry word calculation
Here we are taking a whole file and calculating the slurry word's degree (between 0 and 1), line by line.  Those slurry words will be stored in the list (racing_slurry). So, if any one of the words in the line from the uploaded text document, matches with the word in the list, the count variable will be increased by 1.  Hence the degree is found out by dividing the total number of slurry words by the total words in the line.